requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
Pillow==9.2.0
Pygments==2.13.0
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
